音楽コントロールボタン設置
