+++ src/pages/index.astro
@@ -45,6 +45,7 @@ const formattedDate = `${year}-${month}-${day}`;
                 <li>タグ機能</li>
                 <li>検索機能</li>
                 <li>記事の目次がダサいので、デザインやり直す</li>
+                <li>画面が眩しいのダークテーマ入れる</li>
             </ul>
         </div>
     </div>
